Move language dir into data dir
