+++ data/lang/English/blue_fire_arrow_item_name_eng.ia4
