+++ data/lang/Japanese/blue_fire_arrow_item_name_jap.ia4
