+++ data/lang/Japanese/property.json
